Background_remover/mobile_app/requirements.txt
rembg>=2.0.50
Pillow>=10.0.0
kivy>=2.2.1
plyer>=2.1.0